src/pages/common/components/infobar/Infobar/InfobarCustomerInfo/InfobarWidgets/widgets/magento2/index.ts
import {List, Map} from 'immutable'
import Customer from './Customer'
import Order from './Order'

const magento2 = (args: {
    template: Map<any, any>
    source: Map<any, any>
    parent: Map<any, any>
}) => {
    const path = (args.template.get('absolutePath', []) as List<string>).join(
        '.'
    )

    if (path.match(/integrations\.[0-9]+\.customer$/)) {
        return Customer()
    }

    if (path.match(/integrations\.[0-9]+\.orders.\[]$/)) {
        return Order()
    }

    return {}
}

export default magento2
